requirements.txt
requests
beautifulsoup4
pandas
python-dotenv
openai
jinja2
google-api-python-client
google-auth-httplib2
google-auth-oauthlib